cleaned up cleanup script controller.
