+++ docs-markdown/src/controllers/quick-pick-menu-controller.ts
@@ -5,7 +5,7 @@ import { output } from "../extension";
 import { checkExtension, generateTimestamp } from "../helper/common";
 import { insertAlert } from "./alert-controller";
 import { formatBold } from "./bold-controller";
-import { applyCleanup } from "./cleanup-controller";
+import { applyCleanup } from "./cleanup/cleanup-controller";
 import { formatCode } from "./code-controller";
 import { pickImageType } from "./image-controller";
 import { insertInclude } from "./include-controller";
